Move config map conversion into config.hpp/cpp and remove conversions.cpp
