+++ everestjs/everestjs.cpp
@@ -23,7 +23,6 @@
 #include "js_exec_ctx.hpp"
 #include "utils.hpp"
 
-#include <utils/conversions.hpp>
 #include <utils/error/error_factory.hpp>
 #include <utils/error/error_manager_impl.hpp>
 #include <utils/error/error_manager_req.hpp>
